requirements.txt
ultralytics
toio.py
opencv-python
numpy
asyncio
flask
flask-cors 